Random() change
Creating a new Random() instance each time a random value is needed can cause stutters on lower-end hardware due to Unity's garbage collector needing to clean up those temporary objects. Realistically, most users probably wouldn't notice any performance dips, but it's better to avoid creating those temporary objects in the first place.

Unity has its own built-in random() that doesn't create any temporary instances. Also, UnityEngine.Random is already used in other files (like LethalBotAI.cs and MissionControlState.cs), so this change makes the code more consistant with random usage.

In GetRandomCooldown(bool isResponsiveness), I removed the float cast and made VoicesConst.MIN and VoicesConst.MAX float instead of int. Having them as floats removes the need to cast entirely. Since the GetRandomCooldown(bool isResponsiveness) already returns a float, Random.Range() now returns a decimal value instead of a rounded value. This means the cooldown numbers won't always be round, which should slightly decrease the chances of bots talking right after each other.

The change from System.Random to UnityEngine.Random also doesn't impact the random value generation for the voices. Each call returns its own random value.

This is not the biggest issue or change, but might as well be included with this pull request, as on its own, it's not important or big enough for an update.

+++ Constants/VoicesConst.cs
@@ -15,31 +15,31 @@ namespace LethalBots.Constants
 
         // Talkativeness cooldowns
         // in seconds
-        public const int MIN_COOLDOWN_PLAYVOICE_SHY = 10;
-        public const int MAX_COOLDOWN_PLAYVOICE_SHY = 40;
+        public const float MIN_COOLDOWN_PLAYVOICE_SHY = 10f;
+        public const float MAX_COOLDOWN_PLAYVOICE_SHY = 40f;
 
-        public const int MIN_COOLDOWN_PLAYVOICE_NORMAL = 5;
-        public const int MAX_COOLDOWN_PLAYVOICE_NORMAL = 20;
+        public const float MIN_COOLDOWN_PLAYVOICE_NORMAL = 5f;
+        public const float MAX_COOLDOWN_PLAYVOICE_NORMAL = 20f;
 
-        public const int MIN_COOLDOWN_PLAYVOICE_TALKATIVE = 2;
-        public const int MAX_COOLDOWN_PLAYVOICE_TALKATIVE = 10;
+        public const float MIN_COOLDOWN_PLAYVOICE_TALKATIVE = 2f;
+        public const float MAX_COOLDOWN_PLAYVOICE_TALKATIVE = 10f;
 
-        public const int MIN_COOLDOWN_PLAYVOICE_CANTSTOPTALKING = 0;
-        public const int MAX_COOLDOWN_PLAYVOICE_CANTSTOPTALKING = 0;
+        public const float MIN_COOLDOWN_PLAYVOICE_CANTSTOPTALKING = 0f;
+        public const float MAX_COOLDOWN_PLAYVOICE_CANTSTOPTALKING = 0f;
 
         // Responsiveness cooldowns
         // in seconds
-        public const int MIN_COOLDOWN_PLAYVOICE_RESPONSIVE_SHY = 6;
-        public const int MAX_COOLDOWN_PLAYVOICE_RESPONSIVE_SHY = 12;
+        public const float MIN_COOLDOWN_PLAYVOICE_RESPONSIVE_SHY = 6f;
+        public const float MAX_COOLDOWN_PLAYVOICE_RESPONSIVE_SHY = 12f;
 
-        public const int MIN_COOLDOWN_PLAYVOICE_RESPONSIVE_NORMAL = 2;
-        public const int MAX_COOLDOWN_PLAYVOICE_RESPONSIVE_NORMAL = 6;
+        public const float MIN_COOLDOWN_PLAYVOICE_RESPONSIVE_NORMAL = 2f;
+        public const float MAX_COOLDOWN_PLAYVOICE_RESPONSIVE_NORMAL = 6f;
 
-        public const int MIN_COOLDOWN_PLAYVOICE_RESPONSIVE_RESPONSIVE = 1;
-        public const int MAX_COOLDOWN_PLAYVOICE_RESPONSIVE_RESPONSIVE = 4;
+        public const float MIN_COOLDOWN_PLAYVOICE_RESPONSIVE_RESPONSIVE = 1f;
+        public const float MAX_COOLDOWN_PLAYVOICE_RESPONSIVE_RESPONSIVE = 4f;
 
-        public const int MIN_COOLDOWN_PLAYVOICE_RESPONSIVE_ALWAYSRESPOND = 0;
-        public const int MAX_COOLDOWN_PLAYVOICE_RESPONSIVE_ALWAYSRESPOND = 0;
+        public const float MIN_COOLDOWN_PLAYVOICE_RESPONSIVE_ALWAYSRESPOND = 0f;
+        public const float MAX_COOLDOWN_PLAYVOICE_RESPONSIVE_ALWAYSRESPOND = 0f;
 
         /// <summary>
         /// Voice states that are driven by the responsiveness slider
